# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.GetCategoryDTO:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      category_slug:
        type: string
    type: object
...
  dto.GetProductDTO:
    properties:
      category:
        $ref: '#/definitions/dto.GetCategoryDTO'
      product_description:
        type: string
      product_id:
        type: integer
      product_image:
        type: string
      product_name:
        type: string
      product_price:
        type: number
      seller_id:
        type: integer
    type: object
...
  handler.ErrorResponse:
    properties:
      message:
        example: validation failed
        type: string
    type: object
  handler.MessageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
...
  handler.ProductIDResponse:
    properties:
      product_id:
        example: 15
        type: integer
    type: object
  handler.ProductListResponse:
    properties:
      limit:
        example: 20
        type: integer
      page:
        example: 1
        type: integer
      products:
        items:
          $ref: '#/definitions/dto.GetProductDTO'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 'REST API for marketplace: products, categories, cart, orders, auth
    and roles.'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Marketplace API
  version: "1.0"
paths:
...
  /products:
    get:
      description: Public product listing with search, category filter, sorting and
        pagination.
      parameters:
      - description: Page number (starting from 1)
        in: query
        name: page
        required: true
        type: integer
      - description: Search by product name
        in: query
        name: search
        type: string
      - description: Filter by category slug
        in: query
        name: category
        type: string
      - description: Sort field (e.g. price, name)
        in: query
        name: sortBy
        type: string
      - description: Ascending order (default false)
        in: query
        name: asc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of products
          schema:
            $ref: '#/definitions/handler.ProductListResponse'
        "400":
          description: Invalid page parameter
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List products
      tags:
      - Products
...
  /seller/products:
    get:
      description: List products belonging to the authenticated seller. Requires seller
        or admin role.
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of seller products
          schema:
            $ref: '#/definitions/handler.ProductListResponse'
        "400":
          description: Invalid page
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List seller products
      tags:
      - SellerProducts
    post:
      consumes:
      - multipart/form-data
      description: 'Create a new product with image upload. Requires seller or admin
        role. Rate limit: 5 req/hour. Content-Type: multipart/form-data.'
      parameters:
      - description: Product name
        in: formData
        name: product_name
        required: true
        type: string
      - description: Product description
        in: formData
        name: product_description
        required: true
        type: string
      - description: Product price (decimal)
        in: formData
        name: product_price
        required: true
        type: string
      - description: Category slug
        in: formData
        name: category_slug
        required: true
        type: string
      - description: Product image (image/*)
        in: formData
        name: product_image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Created product ID
          schema:
            $ref: '#/definitions/handler.ProductIDResponse'
        "400":
          description: Parse, validation or missing file error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - SellerProducts
  /seller/products/{product_id}:
    delete:
      description: Delete product by ID. Seller can delete own products, admin can
        delete any. Requires seller or admin role.
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deletion confirmation
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Invalid product_id
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - SellerProducts
    put:
      consumes:
      - multipart/form-data
      description: 'Update existing product. Image is optional. Requires seller or
        admin role. Content-Type: multipart/form-data.'
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      - description: Product name
        in: formData
        name: product_name
        required: true
        type: string
      - description: Product description
        in: formData
        name: product_description
        required: true
        type: string
      - description: Product price (decimal)
        in: formData
        name: product_price
        required: true
        type: string
      - description: Category slug
        in: formData
        name: category_slug
        required: true
        type: string
      - description: New product image (optional)
        in: formData
        name: product_image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Updated product ID
          schema:
            $ref: '#/definitions/handler.ProductIDResponse'
        "400":
          description: Parse or validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - SellerProducts
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"